+++ src/Services/CommentsFetchInfoIntegration/CommentsFetchInfoIntegration.Worker/IntegrationEvents/FetchInfoChangedEvent.cs
@@ -7,9 +7,9 @@ public class FetchInfoChangedEvent : IFetchInfoChangedEvent
     public Guid Id { get; set; }
     public string? VideoId { get; set; }
     public string? PageToken { get; set; }
-    public int CommentsCount { get; set; }
+    public ulong CommentsCount { get; set; }
     public List<string>? CommentIds { get; set; }
-    public int ReplyCount { get; set; }
+    public uint ReplyCount { get; set; }
     public string? Status { get; set; }
     public bool CompletedTillFirstComment { get; set;  }
