+++ src/Services/CommentsFetchInfoIntegration/CommentsFetchInfoIntegration.Worker/Services/Interfaces/IFetchInfoService.cs
@@ -5,7 +5,7 @@ namespace CommentsFetchInfoIntegration.Worker.Services.Interfaces;
 public interface IFetchInfoService
 {
     Task<bool> FetchInfoByIdExistsAsync(string? videoId);
-    Task UpdateFetchInfoAsync(string videoId, int newTotalCommentsFetched, string newStatus);
+    Task UpdateFetchInfoAsync(string videoId, ulong newTotalCommentsFetched, string newStatus);
     Task UpdateFetchInfoAsync(VideoFetchInfo videoFetchInfo);
     Task<VideoFetchInfo> GetFetchInfoByIdAsync(string? videoId);
     Task InsertFetchInfoAsync(VideoFetchInfo videoFetchInfo);
