+++ src/Services/YouTubeCommentsFetcher/YouTubeCommentsFetcher.Worker/IntegrationEvents/CommentThreadListCompletedEvent.cs
@@ -13,8 +13,8 @@ public class CommentThreadListCompletedEvent : IFetchCommentThreadListCompletedE
     public Guid Id { get; }
     public string? VideoId { get; set; }
     public string? NextPageToken { get; set; } 
-    public int CommentsCount { get; set; }
-    public int ReplyCount { get; set; }
+    public ulong CommentsCount { get; set; }
+    public uint ReplyCount { get; set; }
     public List<string>? CommentIds { get; set; }
     public List<IYouTubeCommentDto> YouTubeCommentsList { get; set; }
