+++ src/Services/YouTubeCommentsFetcher/YouTubeCommentsFetcher.Worker/IntegrationEvents/CommentsFetchStatusEvent.cs
@@ -13,9 +13,9 @@ public class FetchInfoChangedEvent : IFetchInfoChangedEvent
     public Guid Id { get; }
     public string VideoId { get; }
     public string? PageToken { get; set; }
-    public int CommentsCount { get; set; }
+    public ulong CommentsCount { get; set; }
     public List<string>? CommentIds { get; set; }
-    public int ReplyCount { get; set; }
+    public uint ReplyCount { get; set; }
     public string? Status { get; set; }
     public bool CompletedTillFirstComment { get; set; }
