+++ src/SharedContracts/IntegrationEventsContracts/IFetchCommentThreadListCompletedEvent.cs
@@ -5,8 +5,8 @@ public interface IFetchCommentThreadListCompletedEvent
     Guid Id { get; }
     string? VideoId { get; }
     string NextPageToken { get; }
-    int CommentsCount { get; }
-    int ReplyCount { get; }
+    ulong CommentsCount { get; }
+    uint ReplyCount { get; }
     List<string>? CommentIds { get; }
     List<IYouTubeCommentDto> YouTubeCommentsList { get; }
 }
